Improve test for Ellen's Alien game
The test is using a zip iterator from the expected input
`position_data` and output, `obj_list`, returned by the
user's new_aliens_collection implementation.

However the zip is not Strict, meaning that if the user provided
a list with less objects than the expected output, the test would
still pass. For example: `return [Alien(x,y)]`.

Add custom assert with an explanatory error about the lenght of the
Alien lsit and call the zip() function with the strict argument
just in case.